Added relationships to my models
I added the specific relationships to the respective models so that I could reference the data and incorporate it into my html.

The data was already available in the User class for us to use.  But we needed to structure or show it how i wanted to be displayed.  I didn't need to use all the data available.