Changed how we present all the user information
Instead of using a rigid table model, we used a ul and li which is a simpler view.

But we had to structure the data into a nested hierarchy because that's how we were recieving the data from the backend.

The real meat of this solution is that were able to show the user's info for all pets, and then for each pet, all prescription information.