added Pet button
I added pet buttons and
added functionality such as:

The meat is that we created a pet using the petService and called Bootstrap data to refresh the BehaviorSubject; and because we were already subscribed to it - it automatically refreshed the page.

The currentUserBehaviorSubject holds and distributes(publishes) the data.
It will notify subscribers of any changes.  And it will notify new subcribers of any changes immediately.
It's the getBootstrapData method that gets(verb) the data by making the http request to the backend and then is able to store it in the currentUserBehaviorSubject.

We added a method to add a pet to the frontend component which would send the newly formed pet information to petService which then adds it to the backend.

We hard-coded a new pet and made sure it worked using console.log.  We subcribed to the addPet method with the new pet parameters through the petService.

Then, we used the subscribe method on the getBootstrapData call of the userService to refresh our page automatically.