moved signUp method
I moved the signUp method/function from the authentication Service to the user Service. This was for better clarity when creating a user - for it to stay with ALL the user functions.

Authentication Service has more to do with creating (requesting), getting, setting, and removing a TOKEN and not necessarily creating a user.  So focus stays with the token in authentication Service.

So the signUp function is better placed within the user Service - for creating a user.