added Update Method
we added an input box to allow the user to input their choice of first name called userFirstName (string) and used [(ngModel)] to data bind to the input string.

we added a button to call the updateCurrentUserFirstName function

when we call the updateCurrentUserFirstName(), this method is found in the userService component which calls the changeFirstName() method which takes 2 arguments: the currentUser's id and the newFirstName (string) that they just typed into the text box.
This method makes an http patch request to the backend to update whatever the user just typed and replace the old value with the new value in the database.

Added 2way data binding to HTML code so that whatever the user types into the text box, it will update in the database.

So now, in the timeline.component.ts, in the updateCurrentUserFirstName(), once this information is sent to the backend, we send an additional request to get the updated information to show to the user automatically on the webpage!