changed pretendToBeBob method from object to string
Template variables help you use data from one part of the template to another.  They can be used to respond to user input.

We orginally set the pretendToBeBob method to send an object to the template variable in the HTML expecting an object structured after our User Model. (see below paragraphs for reiteration and further clarification)

Original HTML code was:
Welcome {{currentUser?.first_name}}
this is refering to a template variable that I use in another part of my component.  It is expecting an object with all parameters specified in the User model.

New HTML code is now:
Welcome <input ... [(ngModel)]="userFirstName">
this is now data binding to the userFirstName (which is defined as a string).  So now, I had to change my code to reflect that.